Snowball.CLI: Refactor Program to Snowball.CLI namespace
